float change to double

+++ catkin_ws/src/AnkerDatasPublish/src/readankerdatafile.cpp
@@ -31,7 +31,7 @@ ReadAnkerDataFile::ReadAnkerDataFile(const string& imu_file,const string& odomet
 
   while(!imu_fin.eof())
   {
-    float t_imu,t_odo,t_opt;
+    double t_imu,t_odo,t_opt;
     imu_fin >> t_imu;
     odo_fin >> t_odo;
     opt_fin >> t_opt;
