+++ catkin_ws/src/anker_extraction/src/ankerdatatype.h
@@ -3,19 +3,19 @@
 
 struct AnkerData
 {
-    float time;
-    float gx;
-    float gy;
-    float gz;
-    float ax;
-    float ay;
-    float az;
-    float odo_right_pos;
-    float odo_left_pos;
-    float odo_right_vel;
-    float odo_left_vel;
-    float opt_pos_x;
-    float opt_pos_y;
+    double time;
+    double gx;
+    double gy;
+    double gz;
+    double ax;
+    double ay;
+    double az;
+    double odo_right_pos;
+    double odo_left_pos;
+    double odo_right_vel;
+    double odo_left_vel;
+    double opt_pos_x;
+    double opt_pos_y;
     unsigned short int opt_quality;
     unsigned short int wall_distance_right;
     unsigned short int wall_distance_left;
